pyproject.toml
[tool.poetry]
name = "project-blog"
version = "0.0.1"
description = "Simple Blog"
authors = ["Dima Cheban"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.34.0"
fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlalchemy = "^2.0.37"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.7.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
alembic = "^1.14.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
